Removed rotation of Coordinates when dragging.
